ref(onboarding): Split Java onboarding docs
