+++ static/app/gettingStartedDocs/java/java/index.tsx
@@ -1,18 +1,16 @@
 import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
-import {feedback} from 'sentry/gettingStartedDocs/java/java/feedback';
 
+import {feedback} from './feedback';
 import {logs} from './logs';
 import {onboarding} from './onboarding';
-import {profiling} from './profiling';
-import {sessionReplay} from './sessionReplay';
+import {platformOptions, type PlatformOptions} from './utils';
 
-const docs: Docs = {
-  onboarding,
+const docs: Docs<PlatformOptions> = {
+  platformOptions,
   feedbackOnboardingCrashApi: feedback,
   crashReportOnboarding: feedback,
-  profilingOnboarding: profiling,
-  replayOnboarding: sessionReplay,
   logsOnboarding: logs,
+  onboarding,
 };
 
 export default docs;
