+++ static/app/gettingStartedDocs/java/java/utils.tsx
@@ -1,0 +1,148 @@
+import type {
+  BasePlatformOptions,
+  DocsParams,
+} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {t} from 'sentry/locale';
+import {getPackageVersion} from 'sentry/utils/gettingStartedDocs/getPackageVersion';
+
+export enum PackageManager {
+  GRADLE = 'gradle',
+  MAVEN = 'maven',
+  SBT = 'sbt',
+}
+
+export enum YesNo {
+  YES = 'yes',
+  NO = 'no',
+}
+
+export const packageManagerName: Record<PackageManager, string> = {
+  [PackageManager.GRADLE]: 'Gradle',
+  [PackageManager.MAVEN]: 'Maven',
+  [PackageManager.SBT]: 'SBT',
+};
+
+export const platformOptions = {
+  packageManager: {
+    label: t('Package Manager'),
+    items: [
+      {
+        label: packageManagerName[PackageManager.GRADLE],
+        value: PackageManager.GRADLE,
+      },
+      {
+        label: packageManagerName[PackageManager.MAVEN],
+        value: PackageManager.MAVEN,
+      },
+      {
+        label: packageManagerName[PackageManager.SBT],
+        value: PackageManager.SBT,
+      },
+    ],
+  },
+  opentelemetry: {
+    label: t('OpenTelemetry'),
+    items: [
+      {
+        label: t('With OpenTelemetry'),
+        value: YesNo.YES,
+      },
+      {
+        label: t('Without OpenTelemetry'),
+        value: YesNo.NO,
+      },
+    ],
+  },
+} satisfies BasePlatformOptions;
+
+export type PlatformOptions = typeof platformOptions;
+export type Params = DocsParams<PlatformOptions>;
+
+export const getGradleInstallSnippet = (params: Params) => `
+buildscript {
+  repositories {
+    mavenCentral()
+  }
+}
+
+plugins {
+  id "io.sentry.jvm.gradle" version "${getPackageVersion(
+    params,
+    'sentry.java.android.gradle-plugin',
+    '3.12.0'
+  )}"
+}
+
+sentry {
+  // Generates a JVM (Java, Kotlin, etc.) source bundle and uploads your source code to Sentry.
+  // This enables source context, allowing you to see your source
+  // code as part of your stack traces in Sentry.
+  includeSourceContext = true
+
+  org = "${params.organization.slug}"
+  projectName = "${params.project.slug}"
+  authToken = System.getenv("SENTRY_AUTH_TOKEN")
+}`;
+
+export const getMavenInstallSnippet = (params: Params) => `
+<build>
+  <plugins>
+    <plugin>
+      <groupId>io.sentry</groupId>
+      <artifactId>sentry-maven-plugin</artifactId>
+      <version>${getPackageVersion(params, 'sentry.java.maven-plugin', '0.0.4')}</version>
+      <extensions>true</extensions>
+      <configuration>
+        <!-- for showing output of sentry-cli -->
+        <debugSentryCli>true</debugSentryCli>
+
+        <org>${params.organization.slug}</org>
+
+        <project>${params.project.slug}</project>
+
+        <!-- in case you're self hosting, provide the URL here -->
+        <!--<url>http://localhost:8000/</url>-->
+
+        <!-- provide your auth token via SENTRY_AUTH_TOKEN environment variable -->
+        <authToken>\${env.SENTRY_AUTH_TOKEN}</authToken>
+      </configuration>
+      <executions>
+        <execution>
+          <goals>
+            <!--
+            Generates a JVM (Java, Kotlin, etc.) source bundle and uploads your source code to Sentry.
+            This enables source context, allowing you to see your source
+            code as part of your stack traces in Sentry.
+            -->
+            <goal>uploadSourceBundle</goal>
+          </goals>
+        </execution>
+      </executions>
+    </plugin>
+  </plugins>
+  ...
+</build>`;
+
+export const getOpenTelemetryRunSnippet = (params: Params) => `
+SENTRY_PROPERTIES_FILE=sentry.properties java -javaagent:sentry-opentelemetry-agent-${getPackageVersion(params, 'sentry.java.opentelemetry-agent', '8.0.0')}.jar -jar your-application.jar
+`;
+
+export const getVerifyJavaSnippet = () => `
+import java.lang.Exception;
+import io.sentry.Sentry;
+
+try {
+  throw new Exception("This is a test.");
+} catch (Exception e) {
+  Sentry.captureException(e);
+}`;
+
+export const getVerifyKotlinSnippet = () => `
+import java.lang.Exception
+import io.sentry.Sentry
+
+try {
+  throw Exception("This is a test.")
+} catch (e: Exception) {
+  Sentry.captureException(e)
+}`;
