+++ static/app/gettingStartedDocs/java/logback/index.tsx
@@ -3,15 +3,13 @@ import {feedback} from 'sentry/gettingStartedDocs/java/java/feedback';
 
 import {logs} from './logs';
 import {onboarding} from './onboarding';
-import {profiling} from './profiling';
-import {sessionReplay} from './sessionReplay';
+import {platformOptions, type PlatformOptions} from './utils';
 
-const docs: Docs = {
+const docs: Docs<PlatformOptions> = {
   onboarding,
   feedbackOnboardingCrashApi: feedback,
   crashReportOnboarding: feedback,
-  profilingOnboarding: profiling,
-  replayOnboarding: sessionReplay,
+  platformOptions,
   logsOnboarding: logs,
 };
