+++ static/app/gettingStartedDocs/java/logback/utils.tsx
@@ -1,0 +1,38 @@
+import type {
+  BasePlatformOptions,
+  DocsParams,
+} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {PackageManager, YesNo} from 'sentry/gettingStartedDocs/java/java/utils';
+import {t} from 'sentry/locale';
+
+export const platformOptions = {
+  packageManager: {
+    label: t('Package Manager'),
+    items: [
+      {
+        label: t('Gradle'),
+        value: PackageManager.GRADLE,
+      },
+      {
+        label: t('Maven'),
+        value: PackageManager.MAVEN,
+      },
+    ],
+  },
+  opentelemetry: {
+    label: t('OpenTelemetry'),
+    items: [
+      {
+        label: t('With OpenTelemetry'),
+        value: YesNo.YES,
+      },
+      {
+        label: t('Without OpenTelemetry'),
+        value: YesNo.NO,
+      },
+    ],
+  },
+} satisfies BasePlatformOptions;
+
+export type PlatformOptions = typeof platformOptions;
+export type Params = DocsParams<PlatformOptions>;
